requirements.txt
asgiref==3.8.1
attrs==25.1.0
certifi==2024.12.14
Django==5.1.5
django-elasticsearch-dsl==8.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
drf-spectacular==0.28.0
elastic-transport==8.17.0
elasticsearch==8.17.1
elasticsearch-dsl==8.17.1
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
#psycopg2==2.9.10
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-dateutil==2.9.0.post0
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.22.3
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.3.0
